+++ ReleaseNotes.md
@@ -6,6 +6,7 @@
     * `hspaced(_: HorizontalAlignment)` and `vspaced(_: VerticalAlignment)`: can be used to wrap a view in an HStack/VStack with Spacer to achieve desired alignment.
 
 ### Improvements
+* SBBHeaderBox: accessibility enhancements, views can be created with `pageContentWithFocus' to have access to the focus state and scroll with the VoiceOver. 
 * SBBRadioButton, SBBCheckBox: new parameter `darkModeCheckWhite`, if enabled the checkmark is white in dark mode, instead of primary color.
 * sbbModal: possibility to give it an `onClose` function to be called when user taps outside the modal view, and therefore closes the modal view.